+++ static/app/views/dashboardsV2/deprecated/detail.spec.jsx
@@ -1,22 +1,16 @@
 import {browserHistory} from 'react-router';
 
+import {createListeners} from 'sentry-test/createListeners';
+import {selectDropdownMenuItem} from 'sentry-test/dropdownMenu';
 import {enforceActOnUseLegacyStoreHook, mountWithTheme} from 'sentry-test/enzyme';
 import {initializeOrg} from 'sentry-test/initializeOrg';
-import {
-  act,
-  render,
-  renderGlobalModal,
-  screen,
-  userEvent,
-  waitFor,
-  within,
-} from 'sentry-test/reactTestingLibrary';
+import {mountGlobalModal} from 'sentry-test/modal';
+import {act} from 'sentry-test/reactTestingLibrary';
+import {triggerPress} from 'sentry-test/utils';
 
 import * as modals from 'sentry/actionCreators/modal';
 import ProjectsStore from 'sentry/stores/projectsStore';
-import CreateDashboard from 'sentry/views/dashboardsV2/create';
-import {constructGridItemKey} from 'sentry/views/dashboardsV2/layoutUtils';
-import {DashboardWidgetSource} from 'sentry/views/dashboardsV2/types';
+import {DashboardState} from 'sentry/views/dashboardsV2/types';
 import * as types from 'sentry/views/dashboardsV2/types';
 import ViewEditDashboard from 'sentry/views/dashboardsV2/view';
 import {OrganizationContext} from 'sentry/views/organizationContext';
@@ -25,18 +19,13 @@ describe('Dashboards > Detail', function () {
   enforceActOnUseLegacyStoreHook();
 
   const organization = TestStubs.Organization({
-    features: [
-      'global-views',
-      'dashboards-basic',
-      'dashboards-edit',
-      'discover-query',
-      'dashboard-grid-layout',
-    ],
+    features: ['global-views', 'dashboards-basic', 'dashboards-edit', 'discover-query'],
   });
   const projects = [TestStubs.Project()];
 
   describe('prebuilt dashboards', function () {
-    let wrapper, initialData;
+    let wrapper;
+    let initialData, mockVisit;
 
     beforeEach(function () {
       act(() => ProjectsStore.loadInitialData(projects));
@@ -61,7 +50,7 @@ describe('Dashboards > Detail', function () {
         url: '/organizations/org-slug/dashboards/default-overview/',
         body: TestStubs.Dashboard([], {id: 'default-overview', title: 'Default'}),
       });
-      MockApiClient.addMockResponse({
+      mockVisit = MockApiClient.addMockResponse({
         url: '/organizations/org-slug/dashboards/1/visit/',
         method: 'POST',
         body: [],
@@ -81,57 +70,105 @@ describe('Dashboards > Detail', function () {
       }
     });
 
-    it('assigns unique IDs to all widgets so grid keys are unique', async function () {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/events-stats/',
-        body: {data: []},
-      });
-      MockApiClient.addMockResponse({
+    it('can delete', async function () {
+      const deleteMock = MockApiClient.addMockResponse({
         url: '/organizations/org-slug/dashboards/default-overview/',
-        body: TestStubs.Dashboard(
-          [
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:error',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'Default Widget 1',
-                interval: '1d',
-              }
-            ),
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:transaction',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'Default Widget 2',
-                interval: '1d',
-              }
-            ),
-          ],
-          {id: 'default-overview', title: 'Default'}
-        ),
+        method: 'DELETE',
       });
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: 'default-overview'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      await tick();
+      wrapper.update();
+
+      // Enter edit mode.
+      wrapper.find('Controls Button[data-test-id="dashboard-edit"]').simulate('click');
+
+      const modal = await mountGlobalModal();
+
+      // Click delete, confirm will show
+      wrapper.find('Controls Button[data-test-id="dashboard-delete"]').simulate('click');
+      await tick();
+
+      await modal.update();
+
+      // Click confirm
+      modal.find('button[aria-label="Confirm"]').simulate('click');
+
+      expect(deleteMock).toHaveBeenCalled();
+    });
+
+    it('can rename and save', async function () {
+      const fireEvent = createListeners('window');
+
+      const updateMock = MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/dashboards/default-overview/',
+        method: 'PUT',
+        body: TestStubs.Dashboard([], {id: '8', title: 'Updated prebuilt'}),
+      });
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: 'default-overview'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      await tick();
+      wrapper.update();
+
+      // Enter edit mode.
+      wrapper.find('Controls Button[data-test-id="dashboard-edit"]').simulate('click');
+
+      await tick();
+      wrapper.update();
+
+      // Rename
+      const dashboardTitle = wrapper.find('DashboardTitle Label');
+      dashboardTitle.simulate('click');
+
+      wrapper.find('StyledInput').simulate('change', {
+        target: {innerText: 'Updated prebuilt', value: 'Updated prebuilt'},
+      });
+
+      act(() => {
+        // Press enter
+        fireEvent.keyDown('Enter');
+      });
+
+      wrapper.find('Controls Button[data-test-id="dashboard-commit"]').simulate('click');
+      await tick();
+      wrapper.update();
+
+      expect(updateMock).toHaveBeenCalledWith(
+        '/organizations/org-slug/dashboards/default-overview/',
+        expect.objectContaining({
+          data: expect.objectContaining({title: 'Updated prebuilt'}),
+        })
+      );
+      // Should redirect to the new dashboard.
+      expect(browserHistory.replace).toHaveBeenCalledWith(
+        expect.objectContaining({
+          pathname: '/organizations/org-slug/dashboard/8/',
+        })
+      );
+    });
+
+    it('disables buttons based on features', async function () {
       initialData = initializeOrg({
         organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'discover-query',
-            'dashboard-grid-layout',
-          ],
+          features: ['global-views', 'dashboards-basic', 'discover-query'],
           projects: [TestStubs.Project()],
         }),
       });
@@ -150,26 +187,22 @@ describe('Dashboards > Detail', function () {
       await tick();
       wrapper.update();
 
-      const dashboardInstance = wrapper.find('Dashboard').instance();
-      const assignedIds = new Set(
-        dashboardInstance.props.dashboard.widgets.map(constructGridItemKey)
-      );
-      expect(assignedIds.size).toBe(dashboardInstance.props.dashboard.widgets.length);
+      // Edit should be disabled
+      const editProps = wrapper
+        .find('Controls Button[data-test-id="dashboard-edit"]')
+        .props();
+      expect(editProps.disabled).toBe(true);
+      expect(mockVisit).not.toHaveBeenCalled();
     });
   });
 
   describe('custom dashboards', function () {
-    let wrapper, initialData, widgets, mockVisit, mockPut;
-
+    let wrapper, initialData, widgets, mockVisit;
     const openEditModal = jest.spyOn(modals, 'openAddDashboardWidgetModal');
+
     beforeEach(function () {
-      window.confirm = jest.fn();
-      initialData = initializeOrg({
-        organization,
-        router: {
-          location: TestStubs.location(),
-        },
-      });
+      initialData = initializeOrg({organization});
+      types.MAX_WIDGETS = 30;
       widgets = [
         TestStubs.Widget(
           [
@@ -177,14 +210,13 @@ describe('Dashboards > Detail', function () {
               name: '',
               conditions: 'event.type:error',
               fields: ['count()'],
-              columns: [],
               aggregates: ['count()'],
+              columns: [],
             },
           ],
           {
             title: 'Errors',
             interval: '1d',
-            widgetType: 'discover',
             id: '1',
           }
         ),
@@ -194,14 +226,13 @@ describe('Dashboards > Detail', function () {
               name: '',
               conditions: 'event.type:transaction',
               fields: ['count()'],
-              columns: [],
               aggregates: ['count()'],
+              columns: [],
             },
           ],
           {
             title: 'Transactions',
             interval: '1d',
-            widgetType: 'discover',
             id: '2',
           }
         ),
@@ -211,8 +242,8 @@ describe('Dashboards > Detail', function () {
               name: '',
               conditions: 'event.type:transaction transaction:/api/cats',
               fields: ['p50()'],
-              columns: [],
               aggregates: ['p50()'],
+              columns: [],
             },
           ],
           {
@@ -252,14 +283,11 @@ describe('Dashboards > Detail', function () {
         ],
       });
       MockApiClient.addMockResponse({
+        method: 'GET',
         url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(widgets, {
-          id: '1',
-          title: 'Custom Errors',
-          filters: {},
-        }),
+        body: TestStubs.Dashboard(widgets, {id: '1', title: 'Custom Errors'}),
       });
-      mockPut = MockApiClient.addMockResponse({
+      MockApiClient.addMockResponse({
         url: '/organizations/org-slug/dashboards/1/',
         method: 'PUT',
         body: TestStubs.Dashboard(widgets, {id: '1', title: 'Custom Errors'}),
@@ -293,10 +321,6 @@ describe('Dashboards > Detail', function () {
         method: 'GET',
         body: [],
       });
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/events-geo/',
-        body: {data: [], meta: {}},
-      });
     });
 
     afterEach(function () {
@@ -304,7 +328,6 @@ describe('Dashboards > Detail', function () {
       jest.clearAllMocks();
       if (wrapper) {
         wrapper.unmount();
-        wrapper = null;
       }
     });
 
@@ -333,6 +356,9 @@ describe('Dashboards > Detail', function () {
       // Enter edit mode.
       wrapper.find('Controls Button[data-test-id="dashboard-edit"]').simulate('click');
 
+      const card = wrapper.find('WidgetCard').first();
+      card.find('WidgetCardPanel').simulate('mouseOver');
+
       // Remove the second and third widgets
       wrapper
         .find('WidgetCard')
@@ -356,7 +382,7 @@ describe('Dashboards > Detail', function () {
         expect.objectContaining({
           data: expect.objectContaining({
             title: 'Custom Errors',
-            widgets: [expect.objectContaining(widgets[0])],
+            widgets: [widgets[0]],
           }),
         })
       );
@@ -365,45 +391,7 @@ describe('Dashboards > Detail', function () {
       expect(mockVisit).toHaveBeenCalledTimes(1);
     });
 
-    it('appends dashboard-level filters to series request', async function () {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(widgets, {
-          id: '1',
-          title: 'Custom Errors',
-          filters: {release: ['abc@1.2.0']},
-        }),
-      });
-      const mock = MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/events-stats/',
-        body: [],
-      });
-
-      wrapper = mountWithTheme(
-        <OrganizationContext.Provider value={initialData.organization}>
-          <ViewEditDashboard
-            organization={initialData.organization}
-            params={{orgId: 'org-slug', dashboardId: '1'}}
-            router={initialData.router}
-            location={initialData.router.location}
-          />
-        </OrganizationContext.Provider>,
-        initialData.routerContext
-      );
-      await tick();
-      wrapper.update();
-
-      expect(mock).toHaveBeenLastCalledWith(
-        '/organizations/org-slug/events-stats/',
-        expect.objectContaining({
-          query: expect.objectContaining({
-            query: 'event.type:transaction transaction:/api/cats release:abc@1.2.0 ',
-          }),
-        })
-      );
-    });
-
-    it('can enter edit mode for widgets', async function () {
+    it('opens edit modal for widgets', async function () {
       wrapper = mountWithTheme(
         <OrganizationContext.Provider value={initialData.organization}>
           <ViewEditDashboard
@@ -438,7 +426,6 @@ describe('Dashboards > Detail', function () {
           widget: {
             id: '1',
             interval: '1d',
-            layout: {h: 2, minH: 2, w: 2, x: 0, y: 0},
             queries: [
               {
                 conditions: 'event.type:error',
@@ -450,9 +437,61 @@ describe('Dashboards > Detail', function () {
             ],
             title: 'Errors',
             type: 'line',
-            widgetType: 'discover',
           },
         })
+      );
+    });
+
+    it('does not update if api update fails', async function () {
+      const fireEvent = createListeners('window');
+      window.confirm = jest.fn(() => true);
+
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/dashboards/1/',
+        method: 'PUT',
+        statusCode: 400,
+      });
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      await tick();
+      wrapper.update();
+
+      // Enter edit mode.
+      wrapper.find('Controls Button[data-test-id="dashboard-edit"]').simulate('click');
+
+      // Rename
+      const dashboardTitle = wrapper.find('DashboardTitle Label');
+      dashboardTitle.simulate('click');
+
+      wrapper.find('StyledInput').simulate('change', {
+        target: {innerText: 'Updated Name', value: 'Updated Name'},
+      });
+
+      act(() => {
+        // Press enter
+        fireEvent.keyDown('Enter');
+      });
+
+      wrapper.find('Controls Button[data-test-id="dashboard-commit"]').simulate('click');
+      await tick();
+      wrapper.update();
+
+      expect(wrapper.find('DashboardTitle EditableText').props().value).toEqual(
+        'Updated Name'
+      );
+      wrapper.find('Controls Button[data-test-id="dashboard-cancel"]').simulate('click');
+
+      expect(wrapper.find('DashboardTitle EditableText').props().value).toEqual(
+        'Custom Errors'
       );
     });
 
@@ -497,44 +536,6 @@ describe('Dashboards > Detail', function () {
       wrapper.update();
       wrapper.find('AddButton[data-test-id="widget-add"]').simulate('click');
       expect(openEditModal).toHaveBeenCalledTimes(1);
-    });
-
-    it('shows top level release filter', async function () {
-      const mockReleases = MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [TestStubs.Release()],
-      });
-
-      initialData = initializeOrg({
-        organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'dashboards-edit',
-            'discover-query',
-            'dashboards-top-level-filter',
-          ],
-          projects: [TestStubs.Project()],
-        }),
-      });
-
-      wrapper = mountWithTheme(
-        <OrganizationContext.Provider value={initialData.organization}>
-          <ViewEditDashboard
-            organization={initialData.organization}
-            params={{orgId: 'org-slug', dashboardId: '1'}}
-            router={initialData.router}
-            location={initialData.router.location}
-          />
-        </OrganizationContext.Provider>,
-        initialData.routerContext
-      );
-      await act(async () => {
-        await tick();
-        wrapper.update();
-      });
-      expect(wrapper.find('ReleasesSelectControl').exists()).toBe(true);
-      expect(mockReleases).toHaveBeenCalledTimes(1);
     });
 
     it('opens widget library when add widget option is clicked', async function () {
@@ -599,773 +600,283 @@ describe('Dashboards > Detail', function () {
       expect(wrapper.find('AddWidget').exists()).toBe(false);
     });
 
-    it('renders successfully if more widgets than stored layouts', async function () {
-      // A case where someone has async added widgets to a dashboard
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(
-          [
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:error',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'First Widget',
-                interval: '1d',
-                id: '1',
-                layout: {i: 'grid-item-1', x: 0, y: 0, w: 2, h: 6},
-              }
-            ),
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:error',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'Second Widget',
-                interval: '1d',
-                id: '2',
-              }
-            ),
+    it('hides and shows breadcrumbs based on feature', async function () {
+      const newOrg = initializeOrg({
+        organization: TestStubs.Organization({
+          features: ['global-views', 'dashboards-basic', 'discover-query'],
+          projects: [TestStubs.Project()],
+        }),
+      });
+
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={newOrg.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={newOrg.router}
+            location={newOrg.router.location}
+          />
+        </OrganizationContext.Provider>,
+        newOrg.routerContext
+      );
+      await tick();
+      wrapper.update();
+
+      expect(wrapper.find('Breadcrumbs').exists()).toBe(false);
+
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      await tick();
+      wrapper.update();
+
+      const breadcrumbs = wrapper.find('Breadcrumbs');
+
+      expect(breadcrumbs.exists()).toBe(true);
+      expect(breadcrumbs.find('BreadcrumbLink').find('a').text()).toEqual('Dashboards');
+      expect(breadcrumbs.find('BreadcrumbItem').last().text()).toEqual('Custom Errors');
+    });
+
+    it('unsets newWidget after rendering', async function () {
+      initialData.router.location = {
+        query: {
+          displayType: 'line',
+          interval: '5m',
+          queryConditions: ['title:test', 'event.type:test'],
+          queryFields: ['count()', 'failure_count()'],
+          queryNames: ['1', '2'],
+          queryOrderby: '',
+          title: 'Widget Title',
+        },
+      };
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={{...initialData.router.location, pathname: '/mockpath'}}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      expect(wrapper.find('DashboardDetail').props().initialState).toEqual(
+        DashboardState.EDIT
+      );
+      expect(wrapper.find('DashboardDetail').props().newWidget).toBeDefined();
+      await act(async () => {
+        // Wrap await tick in act because componentDidMount in Dashboard triggers
+        // state change when parsing widget from location
+        await tick();
+      });
+      wrapper.update();
+
+      // The newWidget state was cleared after adding the widget
+      expect(wrapper.find('DashboardDetail').props().newWidget).toBeUndefined();
+
+      await act(async () => {
+        await tick();
+      });
+    });
+
+    it('enters edit mode when given a new widget in location query', async function () {
+      initialData.router.location = {
+        query: {
+          displayType: 'line',
+          interval: '5m',
+          queryConditions: ['title:test', 'event.type:test'],
+          queryFields: ['count()', 'failure_count()'],
+          queryNames: ['1', '2'],
+          queryOrderby: '',
+          title: 'Widget Title',
+        },
+      };
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={{...initialData.router.location, pathname: '/mockpath'}}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+
+      expect(wrapper.find('DashboardDetail').props().initialState).toEqual(
+        DashboardState.EDIT
+      );
+      await act(async () => {
+        // Wrap await tick in act because componentDidMount in Dashboard triggers
+        // state change when parsing widget from location
+        await tick();
+      });
+    });
+
+    it('enters view mode when not given a new widget in location query', async function () {
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      await tick();
+      wrapper.update();
+      expect(wrapper.find('DashboardDetail').props().initialState).toEqual(
+        DashboardState.VIEW
+      );
+    });
+
+    it('opens add widget to custom modal', async function () {
+      types.MAX_WIDGETS = 10;
+
+      initialData = initializeOrg({
+        organization: TestStubs.Organization({
+          features: [
+            'global-views',
+            'dashboards-basic',
+            'dashboards-edit',
+            'discover-query',
           ],
-          {id: '1', title: 'Custom Errors'}
-        ),
+          projects: [TestStubs.Project()],
+        }),
       });
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={initialData.router}
-          location={initialData.router.location}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
+
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
       );
       await tick();
+      wrapper.update();
 
-      await screen.findByText('First Widget');
-      await screen.findByText('Second Widget');
-    });
+      expect(wrapper.find('Controls Tooltip').prop('disabled')).toBe(true);
 
-    it('does not trigger request if layout not updated', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(
-          [
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:error',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'First Widget',
-                interval: '1d',
-                id: '1',
-                layout: {i: 'grid-item-1', x: 0, y: 0, w: 2, h: 6},
-              }
-            ),
-          ],
-          {id: '1', title: 'Custom Errors'}
-        ),
-      });
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={initialData.router}
-          location={initialData.router.location}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
-      );
-      await tick();
+      // Enter Add Widget mode
+      wrapper
+        .find('Controls Button[data-test-id="add-widget-library"]')
+        .simulate('click');
 
-      userEvent.click(screen.getByText('Edit Dashboard'));
-      userEvent.click(screen.getByText('Save and Finish'));
-      await tick();
-
-      expect(screen.getByText('Edit Dashboard')).toBeInTheDocument();
-      expect(mockPut).not.toHaveBeenCalled();
-    });
-
-    it('renders the custom resize handler for a widget', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(
-          [
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:error',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'First Widget',
-                interval: '1d',
-                id: '1',
-                layout: {i: 'grid-item-1', x: 0, y: 0, w: 2, h: 6},
-              }
-            ),
-          ],
-          {id: '1', title: 'Custom Errors'}
-        ),
-      });
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={initialData.router}
-          location={initialData.router.location}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
-      );
-      await tick();
-
-      userEvent.click(await screen.findByText('Edit Dashboard'));
-      const widget = screen.getByText('First Widget').closest('.react-grid-item');
-      const resizeHandle = within(widget).getByTestId('custom-resize-handle');
-
-      expect(resizeHandle).toBeVisible();
-    });
-
-    it('does not trigger an alert when the widgets have no layout and user cancels without changes', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(
-          [
-            TestStubs.Widget(
-              [
-                {
-                  name: '',
-                  conditions: 'event.type:error',
-                  fields: ['count()'],
-                  aggregates: ['count()'],
-                  columns: [],
-                },
-              ],
-              {
-                title: 'First Widget',
-                interval: '1d',
-                id: '1',
-                layout: null,
-              }
-            ),
-          ],
-          {id: '1', title: 'Custom Errors'}
-        ),
-      });
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={initialData.router}
-          location={initialData.router.location}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
-      );
-      await tick();
-
-      userEvent.click(await screen.findByText('Edit Dashboard'));
-      userEvent.click(await screen.findByText('Cancel'));
-
-      expect(window.confirm).not.toHaveBeenCalled();
-    });
-
-    it('opens the widget viewer modal using the widget id specified in the url', () => {
-      const openWidgetViewerModal = jest.spyOn(modals, 'openWidgetViewerModal');
-      const widget = TestStubs.Widget(
-        [
-          {
-            name: '',
-            conditions: 'event.type:error',
-            fields: ['count()'],
-            aggregates: ['count()'],
-            columns: [],
-            orderby: '',
-          },
-        ],
-        {
-          title: 'First Widget',
-          interval: '1d',
-          id: '1',
-          layout: null,
-        }
-      );
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard([widget], {id: '1', title: 'Custom Errors'}),
-      });
-
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1', widgetId: '1'}}
-          router={initialData.router}
-          location={{...initialData.router.location, pathname: '/widget/123/'}}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
-      );
-
-      expect(openWidgetViewerModal).toHaveBeenCalledWith(
-        expect.objectContaining({
-          organization: initialData.organization,
-          widget,
-          onClose: expect.anything(),
-        })
-      );
-    });
-
-    it('redirects user to dashboard url if widget is not found', () => {
-      const openWidgetViewerModal = jest.spyOn(modals, 'openWidgetViewerModal');
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard([], {id: '1', title: 'Custom Errors'}),
-      });
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1', widgetId: '123'}}
-          router={initialData.router}
-          location={{...initialData.router.location, pathname: '/widget/123/'}}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
-      );
-
-      expect(openWidgetViewerModal).not.toHaveBeenCalled();
-      expect(initialData.router.replace).toHaveBeenCalledWith(
-        expect.objectContaining({
-          pathname: '/organizations/org-slug/dashboard/1/',
-          query: {},
-        })
-      );
-    });
-
-    it('opens the edit widget modal when clicking the edit button', async () => {
-      const widget = TestStubs.Widget(
-        [
-          {
-            name: '',
-            conditions: 'event.type:error',
-            fields: ['count()'],
-            aggregates: ['count()'],
-            columns: [],
-            orderby: '',
-          },
-        ],
-        {
-          title: 'First Widget',
-          interval: '1d',
-          id: '1',
-          layout: null,
-        }
-      );
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard([widget], {id: '1', title: 'Custom Errors'}),
-      });
-
-      render(
-        <ViewEditDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1', widgetId: '1'}}
-          router={initialData.router}
-          location={{...initialData.router.location, pathname: '/widget/123/'}}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
-      );
-
-      renderGlobalModal({context: initialData.routerContext});
-
-      userEvent.click(await screen.findByRole('button', {name: 'Edit Widget'}));
+      expect(openEditModal).toHaveBeenCalledTimes(1);
       expect(openEditModal).toHaveBeenCalledWith(
         expect.objectContaining({
-          widget,
-          organization: initialData.organization,
-          source: DashboardWidgetSource.DASHBOARDS,
+          source: types.DashboardWidgetSource.LIBRARY,
         })
       );
     });
 
-    it('opens the widget viewer modal in a prebuilt dashboard using the widget id specified in the url', () => {
-      const openWidgetViewerModal = jest.spyOn(modals, 'openWidgetViewerModal');
+    it('disables add library widgets when max widgets reached', async function () {
+      types.MAX_WIDGETS = 3;
 
-      render(
-        <CreateDashboard
-          organization={initialData.organization}
-          params={{orgId: 'org-slug', templateId: 'default-template', widgetId: '2'}}
-          router={initialData.router}
-          location={{...initialData.router.location, pathname: '/widget/2/'}}
-        />,
-        {context: initialData.routerContext, organization: initialData.organization}
+      initialData = initializeOrg({
+        organization: TestStubs.Organization({
+          features: [
+            'global-views',
+            'dashboards-basic',
+            'dashboards-edit',
+            'discover-query',
+          ],
+          projects: [TestStubs.Project()],
+        }),
+      });
+
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
       );
+      await tick();
+      wrapper.update();
 
-      expect(openWidgetViewerModal).toHaveBeenCalledWith(
+      expect(wrapper.find('WidgetCard')).toHaveLength(3);
+      expect(
+        wrapper.find('Controls Button[data-test-id="add-widget-library"]').props()
+          .disabled
+      ).toEqual(true);
+      expect(wrapper.find('Controls Tooltip').prop('disabled')).toBe(false);
+
+      await act(async () => {
+        triggerPress(wrapper.first().find('MenuControlWrap Button').first());
+
+        await tick();
+        wrapper.update();
+      });
+
+      expect(
+        wrapper.find(`MenuItemWrap[data-test-id="duplicate-widget"]`).props()[
+          'aria-disabled'
+        ]
+      ).toEqual(true);
+    });
+
+    it('opens edit modal when editing widget from context menu', async function () {
+      wrapper = mountWithTheme(
+        <OrganizationContext.Provider value={initialData.organization}>
+          <ViewEditDashboard
+            organization={initialData.organization}
+            params={{orgId: 'org-slug', dashboardId: '1'}}
+            router={initialData.router}
+            location={initialData.router.location}
+          />
+        </OrganizationContext.Provider>,
+        initialData.routerContext
+      );
+      await tick();
+      wrapper.update();
+
+      expect(wrapper.find('WidgetCard')).toHaveLength(3);
+
+      await selectDropdownMenuItem({
+        wrapper,
+        specifiers: {prefix: 'WidgetCard', first: true},
+        itemKey: 'edit-widget',
+      });
+
+      expect(openEditModal).toHaveBeenCalledTimes(1);
+      expect(openEditModal).toHaveBeenCalledWith(
         expect.objectContaining({
-          organization: initialData.organization,
-          widget: expect.objectContaining({
-            displayType: 'line',
-            interval: '5m',
+          widget: {
+            id: '1',
+            interval: '1d',
             queries: [
               {
+                conditions: 'event.type:error',
+                fields: ['count()'],
                 aggregates: ['count()'],
                 columns: [],
-                conditions: '!event.type:transaction',
-                fields: ['count()'],
-                name: 'Events',
-                orderby: 'count()',
+                name: '',
               },
             ],
-            title: 'Events',
-            widgetType: 'discover',
-          }),
-          onClose: expect.anything(),
-        })
-      );
-    });
-
-    it('saves a new dashboard with the page filters', async () => {
-      const mockPOST = MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/',
-        method: 'POST',
-        body: [],
-      });
-      render(
-        <CreateDashboard
-          organization={{
-            ...initialData.organization,
-            features: [
-              ...initialData.organization.features,
-              'dashboards-top-level-filter',
-            ],
-          }}
-          params={{orgId: 'org-slug'}}
-          router={initialData.router}
-          location={{
-            ...initialData.router.location,
-            query: {
-              ...initialData.router.location.query,
-              statsPeriod: '7d',
-              project: [2],
-              environment: ['alpha', 'beta'],
-            },
-          }}
-        />,
-        {
-          context: initialData.routerContext,
-          organization: initialData.organization,
-        }
-      );
-
-      userEvent.click(await screen.findByText('Save and Finish'));
-      expect(mockPOST).toHaveBeenCalledWith(
-        '/organizations/org-slug/dashboards/',
-        expect.objectContaining({
-          data: expect.objectContaining({
-            projects: [2],
-            environment: ['alpha', 'beta'],
-            period: '7d',
-          }),
-        })
-      );
-    });
-
-    it('saves a template with the page filters', async () => {
-      const mockPOST = MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/',
-        method: 'POST',
-        body: [],
-      });
-      render(
-        <CreateDashboard
-          organization={{
-            ...initialData.organization,
-            features: [
-              ...initialData.organization.features,
-              'dashboards-top-level-filter',
-            ],
-          }}
-          params={{orgId: 'org-slug', templateId: 'default-template'}}
-          router={initialData.router}
-          location={{
-            ...initialData.router.location,
-            query: {
-              ...initialData.router.location.query,
-              statsPeriod: '7d',
-              project: [2],
-              environment: ['alpha', 'beta'],
-            },
-          }}
-        />,
-        {
-          context: initialData.routerContext,
-          organization: initialData.organization,
-        }
-      );
-
-      userEvent.click(await screen.findByText('Add Dashboard'));
-      expect(mockPOST).toHaveBeenCalledWith(
-        '/organizations/org-slug/dashboards/',
-        expect.objectContaining({
-          data: expect.objectContaining({
-            projects: [2],
-            environment: ['alpha', 'beta'],
-            period: '7d',
-          }),
-        })
-      );
-    });
-
-    it('does not render save and cancel buttons on templates', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [
-          TestStubs.Release({
-            shortVersion: 'sentry-android-shop@1.2.0',
-            version: 'sentry-android-shop@1.2.0',
-          }),
-        ],
-      });
-      render(
-        <CreateDashboard
-          organization={{
-            ...initialData.organization,
-            features: [
-              ...initialData.organization.features,
-              'dashboards-top-level-filter',
-            ],
-          }}
-          params={{orgId: 'org-slug', templateId: 'default-template'}}
-          router={initialData.router}
-          location={initialData.router.location}
-        />,
-        {
-          context: initialData.routerContext,
-          organization: initialData.organization,
-        }
-      );
-
-      userEvent.click(await screen.findByText('24H'));
-      userEvent.click(screen.getByText('Last 7 days'));
-
-      expect(screen.queryByText('Cancel')).not.toBeInTheDocument();
-      expect(screen.queryByText('Save')).not.toBeInTheDocument();
-    });
-
-    it('can save dashboard filters in existing dashboard', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [
-          TestStubs.Release({
-            shortVersion: 'sentry-android-shop@1.2.0',
-            version: 'sentry-android-shop@1.2.0',
-          }),
-        ],
-      });
-      const testData = initializeOrg({
-        organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'dashboards-edit',
-            'discover-query',
-            'dashboard-grid-layout',
-            'dashboards-top-level-filter',
-          ],
-        }),
-        router: {
-          location: {
-            ...TestStubs.location(),
-            query: {
-              statsPeriod: '7d',
-            },
+            title: 'Errors',
+            type: 'line',
           },
-        },
-      });
-      render(
-        <ViewEditDashboard
-          organization={testData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={testData.router}
-          location={testData.router.location}
-        />,
-        {context: testData.routerContext, organization: testData.organization}
-      );
-
-      await screen.findByText('7D');
-      userEvent.click(await screen.findByText('All Releases'));
-      userEvent.click(screen.getByText('sentry-android-shop@1.2.0'));
-
-      userEvent.click(screen.getByText('Save'));
-
-      expect(mockPut).toHaveBeenCalledWith(
-        '/organizations/org-slug/dashboards/1/',
-        expect.objectContaining({
-          data: expect.objectContaining({
-            period: '7d',
-            filters: {release: ['sentry-android-shop@1.2.0']},
-          }),
         })
       );
-    });
-
-    it('can save absolute time range in existing dashboard', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [],
-      });
-      const testData = initializeOrg({
-        organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'dashboards-edit',
-            'discover-query',
-            'dashboard-grid-layout',
-            'dashboards-top-level-filter',
-          ],
-        }),
-        router: {
-          location: {
-            ...TestStubs.location(),
-            query: {
-              start: '2022-07-14T07:00:00',
-              end: '2022-07-19T23:59:59',
-              utc: 'true',
-            },
-          },
-        },
-      });
-      render(
-        <ViewEditDashboard
-          organization={testData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={testData.router}
-          location={testData.router.location}
-        />,
-        {context: testData.routerContext, organization: testData.organization}
-      );
-
-      userEvent.click(await screen.findByText('Save'));
-
-      expect(mockPut).toHaveBeenCalledWith(
-        '/organizations/org-slug/dashboards/1/',
-        expect.objectContaining({
-          data: expect.objectContaining({
-            start: '2022-07-14T07:00:00.000',
-            end: '2022-07-19T23:59:59.000',
-            utc: true,
-          }),
-        })
-      );
-    });
-
-    it('can clear dashboard filters in existing dashboard', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [
-          TestStubs.Release({
-            shortVersion: 'sentry-android-shop@1.2.0',
-            version: 'sentry-android-shop@1.2.0',
-          }),
-        ],
-      });
-      const testData = initializeOrg({
-        organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'dashboards-edit',
-            'discover-query',
-            'dashboard-grid-layout',
-            'dashboards-top-level-filter',
-          ],
-        }),
-        router: {
-          location: {
-            ...TestStubs.location(),
-            query: {
-              statsPeriod: '7d',
-              project: [1, 2],
-              environment: ['alpha', 'beta'],
-            },
-          },
-        },
-      });
-      render(
-        <ViewEditDashboard
-          organization={testData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={testData.router}
-          location={testData.router.location}
-        />,
-        {context: testData.routerContext, organization: testData.organization}
-      );
-
-      await screen.findByText('7D');
-      userEvent.click(await screen.findByText('All Releases'));
-      userEvent.click(screen.getByText('sentry-android-shop@1.2.0'));
-      userEvent.keyboard('{esc}');
-
-      userEvent.click(screen.getByText('Cancel'));
-
-      screen.getByText('All Releases');
-      expect(browserHistory.replace).toHaveBeenCalledWith(
-        expect.objectContaining({
-          query: expect.objectContaining({
-            project: undefined,
-            statsPeriod: undefined,
-            environment: undefined,
-          }),
-        })
-      );
-    });
-
-    it('disables dashboard actions when there are unsaved filters', async () => {
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [
-          TestStubs.Release({
-            shortVersion: 'sentry-android-shop@1.2.0',
-            version: 'sentry-android-shop@1.2.0',
-          }),
-        ],
-      });
-      const testData = initializeOrg({
-        organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'dashboards-edit',
-            'discover-basic',
-            'discover-query',
-            'dashboard-grid-layout',
-            'dashboards-top-level-filter',
-          ],
-        }),
-        router: {
-          location: {
-            ...TestStubs.location(),
-            query: {
-              statsPeriod: '7d',
-              project: [1, 2],
-              environment: ['alpha', 'beta'],
-            },
-          },
-        },
-      });
-      render(
-        <ViewEditDashboard
-          organization={testData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={testData.router}
-          location={testData.router.location}
-        />,
-        {context: testData.routerContext, organization: testData.organization}
-      );
-
-      expect(await screen.findByText('Save')).toBeInTheDocument();
-      expect(screen.getByText('Cancel')).toBeInTheDocument();
-      expect(screen.getByRole('button', {name: 'Add Widget'})).toBeDisabled();
-      expect(screen.getByRole('button', {name: 'Edit Dashboard'})).toBeDisabled();
-
-      userEvent.click(screen.getAllByLabelText('Widget actions')[0]);
-
-      expect(screen.getByTestId('edit-widget')).toHaveAttribute('aria-disabled', 'true');
-      expect(screen.getByTestId('duplicate-widget')).toHaveAttribute(
-        'aria-disabled',
-        'true'
-      );
-      expect(screen.getByTestId('delete-widget')).toHaveAttribute(
-        'aria-disabled',
-        'true'
-      );
-
-      // Open in discover shouldn't be disabled
-      expect(screen.getByTestId('open-in-discover')).toHaveAttribute(
-        'aria-disabled',
-        'false'
-      );
-    });
-
-    it('ignores the order of selection of page filters to render unsaved filters', async () => {
-      const testProjects = [
-        TestStubs.Project({id: '1', name: 'first', environments: ['alpha', 'beta']}),
-        TestStubs.Project({id: '2', name: 'second', environments: ['alpha', 'beta']}),
-      ];
-
-      act(() => ProjectsStore.loadInitialData(testProjects));
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/projects/',
-        body: testProjects,
-      });
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/releases/',
-        body: [],
-      });
-
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/dashboards/1/',
-        body: TestStubs.Dashboard(widgets, {
-          id: '1',
-          title: 'Custom Errors',
-          filters: {},
-          environment: ['alpha', 'beta'],
-        }),
-      });
-      const testData = initializeOrg({
-        organization: TestStubs.Organization({
-          features: [
-            'global-views',
-            'dashboards-basic',
-            'dashboards-edit',
-            'discover-query',
-            'dashboard-grid-layout',
-            'dashboards-top-level-filter',
-          ],
-        }),
-        router: {
-          location: {
-            ...TestStubs.location(),
-            query: {
-              environment: ['beta', 'alpha'], // Reversed order from saved dashboard
-            },
-          },
-        },
-      });
-      render(
-        <ViewEditDashboard
-          organization={testData.organization}
-          params={{orgId: 'org-slug', dashboardId: '1'}}
-          router={testData.router}
-          location={testData.router.location}
-        />,
-        {context: testData.routerContext, organization: testData.organization}
-      );
-
-      await waitFor(() => expect(screen.queryAllByText('Loading\u2026')).toEqual([]));
-      await screen.findByText(/beta, alpha/);
-
-      // Save and Cancel should not appear because alpha, beta is the same as beta, alpha
-      expect(screen.queryByText('Save')).not.toBeInTheDocument();
-      expect(screen.queryByText('Cancel')).not.toBeInTheDocument();
     });
   });
 });
